Remove suporte a tópicos de retry e DLT, reestrutura reprocessamento de pagamentos pendentes, aprimora resiliência e ajusta logs

+++ services/payment/src/main/java/com/connectfood/payment/infrastructure/http/dto/ProcpagRequest.java
@@ -2,10 +2,8 @@ package com.connectfood.payment.infrastructure.http.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import java.math.BigDecimal;
-
 public record ProcpagRequest(
-    @JsonProperty("valor") BigDecimal valor,
+    @JsonProperty("valor") Long valor,
     @JsonProperty("pagamento_id") String pagamentoId,
     @JsonProperty("cliente_id") String clienteId
 ) {
